Normalize shared layout across pages
