+++ src/components/layout/Footer.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import FooterBase from "../Footer";
+
+export default function Footer() {
+  return <FooterBase />;
+}
