+++ src/components/layout/Header.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Navbar from "../Navbar";
+
+export default function Header() {
+  return (
+    <header className="border-b border-slate-800 bg-[#050B1A]/80 backdrop-blur">
+      <Navbar />
+    </header>
+  );
+}
