+++ src/components/layout/Layout.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Header from "./Header";
+import Footer from "./Footer";
+
+export default function Layout({ children }) {
+  return (
+    <div className="min-h-screen flex flex-col bg-[#050B1A] text-slate-100">
+      <Header />
+      <main className="flex-1">{children}</main>
+      <Footer />
+    </div>
+  );
+}
