+++ src/pages/HowItWorks.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import PageShell from "../components/layout/PageShell";
 import MarketSentimentVisualizer from "../components/ai/MarketSentimentVisualizer";
 
 const workflowSteps = [
@@ -69,8 +68,7 @@ const faqItems = [
 
 export default function HowItWorks() {
   return (
-    <PageShell>
-      <div className="bg-[#050B1A] text-slate-100 min-h-screen">
+    <div className="bg-[#050B1A] text-slate-100 min-h-screen">
         <section className="text-center py-12">
           <div className="relative overflow-hidden">
             <div className="absolute -top-20 right-10 hidden h-72 w-72 rounded-full bg-emerald-500/10 blur-3xl lg:block" />
@@ -184,6 +182,5 @@ export default function HowItWorks() {
           </div>
         </section>
       </div>
-    </PageShell>
   );
 }
